detect/template: clean up packet keyword
